--- src/form/ko/index.ts
@@ -1,3 +1,0 @@
-export * from './formEditor';
-export * from './formViewModel';
-export * from './formViewModelBinder';
