+++ src/forms.module.ts
@@ -1,10 +1,10 @@
 import { IInjectorModule, IInjector } from "@paperbits/common/injection";
 import { IModelBinder, IWidgetHandler } from "@paperbits/common/editing";
 import { IViewModelBinder } from "@paperbits/common/widgets";
-import { FormModule } from "./form/form.module";
-import { InputModule } from "./input/input.module";
-import { SelectModule } from "./select/select.module";
-import { TextareaModule } from "./textarea/textarea.module";
+import { FormModule } from "./form/ko/form.module";
+import { InputModule } from "./input/ko/input.module";
+import { SelectModule } from "./select/ko/select.module";
+import { TextareaModule } from "./textarea/ko/textarea.module";
 
 export class FormsModule implements IInjectorModule {
     constructor(
