--- src/input/ko/index.ts
@@ -1,3 +1,0 @@
-export * from './inputEditor';
-export * from './inputViewModel';
-export * from './inputViewModelBinder';
