--- src/select/ko/index.ts
@@ -1,3 +1,0 @@
-export * from './selectEditor';
-export * from './selectViewModel';
-export * from './selectViewModelBinder';
