--- src/textarea/ko/index.ts
@@ -1,3 +1,0 @@
-export * from './textareaEditor';
-export * from './textareaViewModel';
-export * from './textareaViewModelBinder';
